Move AnnotationExcluder to psi-utils
