+++ detekt-api/api/detekt-api.api
@@ -2,11 +2,6 @@ public abstract interface annotation class io/gitlab/arturbosch/detekt/api/Activ
 	public abstract fun since ()Ljava/lang/String;
 }
 
-public final class io/gitlab/arturbosch/detekt/api/AnnotationExcluder {
-	public fun <init> (Lorg/jetbrains/kotlin/psi/KtFile;Ljava/util/List;Lorg/jetbrains/kotlin/resolve/BindingContext;)V
-	public final fun shouldExclude (Ljava/util/List;)Z
-}
-
 public class io/gitlab/arturbosch/detekt/api/CodeSmell : io/gitlab/arturbosch/detekt/api/Finding {
 	public fun <init> (Lio/gitlab/arturbosch/detekt/api/Entity;Ljava/lang/String;Ljava/util/List;)V
 	public synthetic fun <init> (Lio/gitlab/arturbosch/detekt/api/Entity;Ljava/lang/String;Ljava/util/List;ILkotlin/jvm/internal/DefaultConstructorMarker;)V
