+++ detekt-psi-utils/src/test/kotlin/io/github/detekt/psi/AnnotationExcluderSpec.kt
@@ -1,4 +1,4 @@
-package io.gitlab.arturbosch.detekt.api
+package io.github.detekt.psi
 
 import io.github.detekt.test.utils.compileContentForTest
 import io.gitlab.arturbosch.detekt.rules.KotlinCoreEnvironmentTest
